remove redundancy
