+++ .editorconfig
@@ -6,6 +6,7 @@ insert_final_newline = true
 max_line_length = 120
 
 [*.java]
+max_line_length = 140
 ij_continuation_indent_size = 4
 ij_java_class_count_to_use_import_on_demand = 999
 ij_java_imports_layout = $*, |, java.**, |, javax.**, |, org.**, |, net.**, |, com.**, |, *
