+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/bestpractices/AccessorMethodGenerationRule.java
@@ -4,9 +4,7 @@
 
 package net.sourceforge.pmd.lang.java.rule.bestpractices;
 
-import java.lang.reflect.Modifier;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 import net.sourceforge.pmd.lang.java.ast.ASTExpression;
@@ -15,11 +13,9 @@ import net.sourceforge.pmd.lang.java.ast.ASTMethodCall;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableAccess;
 import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRulechainRule;
+import net.sourceforge.pmd.lang.java.rule.bestpractices.accessor.AccessorHelper;
 import net.sourceforge.pmd.lang.java.symbols.JAccessibleElementSymbol;
-import net.sourceforge.pmd.lang.java.symbols.JClassSymbol;
-import net.sourceforge.pmd.lang.java.symbols.JConstructorSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JFieldSymbol;
-import net.sourceforge.pmd.lang.java.symbols.JMethodSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JVariableSymbol;
 import net.sourceforge.pmd.reporting.RuleContext;
 
@@ -48,57 +44,24 @@ public class AccessorMethodGenerationRule extends AbstractJavaRulechainRule {
 
     @Override
     public Object visit(ASTVariableAccess node, Object data) {
-        JVariableSymbol sym = node.getReferencedSym();
-        if (sym instanceof JFieldSymbol) {
-            JFieldSymbol fieldSym = (JFieldSymbol) sym;
-            if (((JFieldSymbol) sym).getConstValue() == null) {
-                checkMemberAccess((RuleContext) data, node, fieldSym);
-            }
-        }
+        visit(node, (RuleContext) data, node.getReferencedSym());
         return null;
+    }
+
+    private void visit(ASTVariableAccess node, RuleContext data, JVariableSymbol sym) {
+        if (sym instanceof JFieldSymbol && ((JFieldSymbol) sym).getConstValue() == null) {
+            checkMemberAccess(data, node, (JFieldSymbol) sym);
+        }
     }
 
     @Override
     public Object visit(ASTMethodCall node, Object data) {
-        JMethodSymbol symbol = (JMethodSymbol) node.getMethodType().getSymbol();
-        checkMemberAccess((RuleContext) data, node, symbol);
+        checkMemberAccess((RuleContext) data, node, node.getMethodType().getSymbol());
         return null;
     }
 
     private void checkMemberAccess(RuleContext data, ASTExpression node, JAccessibleElementSymbol symbol) {
-        checkMemberAccess(data, node, symbol, this.reportedNodes);
+        AccessorHelper.checkMemberAccess(data, node, symbol, reportedNodes);
     }
 
-    static void checkMemberAccess(RuleContext ruleContext, JavaNode refExpr, JAccessibleElementSymbol sym, Set<JavaNode> reportedNodes) {
-        if (Modifier.isPrivate(sym.getModifiers())
-            && !Objects.equals(sym.getEnclosingClass(),
-                               refExpr.getEnclosingType().getSymbol())) {
-
-            JavaNode node = sym.tryGetNode();
-            if (node == null && JConstructorSymbol.CTOR_NAME.equals(sym.getSimpleName())) {
-                // might be a default constructor, implicitly defined and not explicitly in the compilation unit
-                node = sym.getEnclosingClass().tryGetNode();
-            }
-            assert node != null : "Node should be in the same compilation unit";
-            if (reportedNodes.add(node)) {
-                ruleContext.addViolation(node, stripPackageName(refExpr.getEnclosingType().getSymbol()));
-            }
-        }
-    }
-
-    /**
-     * Returns the canonical name without the package name. Eg for a
-     * canonical name {@code com.github.Outer.Inner}, returns {@code Outer.Inner}.
-     */
-    private static String stripPackageName(JClassSymbol symbol) {
-        String p = symbol.getPackageName();
-        String canoName = symbol.getCanonicalName();
-        if (canoName == null) {
-            return symbol.getSimpleName();
-        }
-        if (p.isEmpty()) {
-            return canoName;
-        }
-        return canoName.substring(p.length() + 1); //+1 for the dot
-    }
 }
